requirements.txt

grpcio==1.15.0
opentracing==1.3.0
six==1.10.0
setuptools>=39.0.1
jaeger-client==3.11.0